requirements-dev.txt
# Development Dependencies

# Code formatting
black>=23.0.0

# Linting
pylint>=3.0.0
flake8>=6.0.0

# Type checking
mypy>=1.5.0

# Test coverage
pytest-cov>=4.1.0
